backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
mediapipe==0.10.14
numpy==1.24.4
scipy==1.11.4
imutils==0.5.4
pygame==2.5.2
python-dotenv==1.0.0
